feature/rename-organisation
Changes
----
* Updated: package name so that country is the first folder
* Updated: implicit import required for Play JSON AST <-> object made specific
