--- play-json/src/test/scala/org/aeb/uk/sandbox/play/PlayJsonSpec.scala
@@ -1,60 +1,0 @@
-package org.aeb.uk.sandbox.play
-
-import org.aeb.uk.sandbox.play.models.{Address, Person}
-import org.scalatest.flatspec.AnyFlatSpec
-import org.scalatest.matchers.should
-import play.api.libs.json._
-
-class PlayJsonSpec extends AnyFlatSpec with should.Matchers {
-
-  val AddressJson = """{"city":"Los Angeles","postCode":"Unknown","country":"USA","house":"Unknown","region":"California","street":"Unknown"}"""
-  val AddressObject: Address = Address(city="Los Angeles", country="USA", house="Unknown", postCode="Unknown", region="California", street="Unknown")
-  val AddressAST: JsValue = JsObject(
-    Map(
-      "house" -> JsString("Unknown"),
-      "street" -> JsString("Unknown"),
-      "city" -> JsString("Los Angeles"),
-      "region" -> JsString("California"),
-      "country" -> JsString("USA"),
-      "postCode" -> JsString("Unknown")
-    )
-  )
-  val PersonJson = s"""{"firstName":"Tyler","lastName":"Durden","age":45,"address":$AddressJson}"""
-  val PersonObject: Person = Person(firstName="Tyler", lastName="Durden", age=45, address=AddressObject)
-  val PersonAST: JsValue = JsObject(
-    Map(
-      "firstName" -> JsString("Tyler"),
-      "lastName" -> JsString("Durden"),
-      "age" -> JsNumber(45),
-      "address" -> AddressAST)
-  )
-
-  "Play JSON" should "convert a JSON string to its Abstract Syntax Tree (AST) representation" in {
-
-    Json.parse(PersonJson) shouldBe PersonAST
-
-  }
-
-  it should "convert an AST to a JSON string" in {
-
-    Json.stringify(PersonAST) shouldBe PersonJson
-
-  }
-
-  it should "convert a class instance to an AST" in {
-
-    import org.aeb.uk.sandbox.play.models.Person._
-
-    Json.toJson(PersonObject) shouldBe PersonAST
-
-  }
-
-  it should "convert an AST to a class instance" in {
-
-    import org.aeb.uk.sandbox.play.models.Person._
-
-    Json.fromJson(PersonAST) shouldBe JsSuccess(PersonObject)
-
-  }
-
-}
